requirements.txt
xgboost
watchdog
streamlit
scikit-learn
python-dotenv
plotly
openai
requests
supabase
Pillow